Split apiserver flags and initialization
Make it easier to keep defaults and flags clearly identified.
