+++ cmd/hyperkube/kube-apiserver.go
@@ -19,19 +19,20 @@ limitations under the License.
 package main
 
 import (
-	kubeapiserver "k8s.io/kubernetes/cmd/kube-apiserver/app"
+	"k8s.io/kubernetes/cmd/kube-apiserver/app"
+	"k8s.io/kubernetes/cmd/kube-apiserver/app/options"
 )
 
 // NewKubeAPIServer creates a new hyperkube Server object that includes the
 // description and flags.
 func NewKubeAPIServer() *Server {
-	s := kubeapiserver.NewAPIServer()
+	s := options.NewAPIServer()
 
 	hks := Server{
 		SimpleUsage: "apiserver",
 		Long:        "The main API entrypoint and interface to the storage system.  The API server is also the focal point for all authorization decisions.",
 		Run: func(_ *Server, args []string) error {
-			return s.Run(args)
+			return app.Run(s)
 		},
 	}
 	s.AddFlags(hks.Flags())
