+++ cmd/kube-apiserver/app/server_test.go
@@ -22,6 +22,7 @@ import (
 	"strings"
 	"testing"
 
+	"k8s.io/kubernetes/cmd/kube-apiserver/app/options"
 	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/api/meta"
 	"k8s.io/kubernetes/pkg/apis/extensions"
@@ -30,7 +31,7 @@ import (
 )
 
 func TestLongRunningRequestRegexp(t *testing.T) {
-	regexp := regexp.MustCompile(defaultLongRunningRequestRE)
+	regexp := regexp.MustCompile(options.NewAPIServer().LongRunningRequestRE)
 	dontMatch := []string{
 		"/api/v1/watch-namespace/",
 		"/api/v1/namespace-proxy/",
@@ -232,10 +233,10 @@ func TestParseRuntimeConfig(t *testing.T) {
 		},
 	}
 	for _, test := range testCases {
-		s := &APIServer{
+		s := &options.APIServer{
 			RuntimeConfig: test.runtimeConfig,
 		}
-		apiGroupVersionOverrides, err := s.parseRuntimeConfig()
+		apiGroupVersionOverrides, err := parseRuntimeConfig(s)
 
 		if err == nil && test.err {
 			t.Fatalf("expected error for test: %q", test)
